refactor: delete rollout.py, inline collect_sync_batch into train.py
rollout.py only had collect_sync_batch + RolloutStats left. Inline them
into train.py where they're actually used (by run_rl_loop). No more
hidden rollout logic in utils — the RL flow is fully visible in the
loop files.
